execl*(), fixed /dev/fb, signal functions, calloc(), misc faster rand() proper calloc() concurrent file accesses redone vfs readdir fixed bug on closing of pipes that would lead to multiple processes 	being stuck in sleep waiting on pipe read proper waitpid() when sigchld is blocked for WCONTINUED and WSTOPPED fixed rw locks leaking critical section counters fixed TOSTOP being ignored in TTY fixed multiple concurency and locking bugs in setpgid fixed setpgid to act more POSIX-y

+++ src/fs/inode.c
@@ -156,7 +156,7 @@ void close_inode(inode_t *inode) {
 
 long inode_from_device(dev_t device, inode_t ** inode_out) {
     if (inode_out == NULL) return -EFAULT;
-    static unsigned long long ephemeral_id = 0;
+    static unsigned long ephemeral_id = 0;
 
     inode_t new = {
         .id = (void*)__atomic_fetch_add(&ephemeral_id, 1, __ATOMIC_RELAXED),
